Bank Marketing Campaigns
The data is related with direct bank marketing campaigns( phone call). marketing refers to activities undertaken by a company to promote the buying or selling of a product or service. Marketing includes advertising , selling, and delivering products to consumers or other bussinesses.
The classification goal is to predict if the client will subscribe a term deposit.